Split regular RepeatTestSuite and PhptRepeatTestSuite
